day 5 part 1 - make everything Long

+++ src/main/kotlin/com/oocode/Almanac.kt
@@ -2,7 +2,7 @@ package com.oocode
 
 fun almanacFrom(input: String): Almanac {
     val lines = input.split("\n")
-    val seeds = lines[0].split(" ").drop(1).map { it.toInt() }
+    val seeds = lines[0].split(" ").drop(1).map { it.toLong() }
 
     val converters = mutableListOf<Converter>()
     var currentMappings = mutableListOf<Mapping>()
@@ -25,21 +25,21 @@ fun almanacFrom(input: String): Almanac {
 private fun numbersFrom(line: String) =
     Regex("(\\d+)")
         .findAll(line)
-        .map { it.value.toInt() }
+        .map { it.value.toLong() }
         .toList()
 
-class Almanac(private val seeds: List<Int>, private val converterChain: ConverterChain) {
+class Almanac(private val seeds: List<Long>, private val converterChain: ConverterChain) {
     fun lowestLocationNumber() = seeds.minOf { converterChain.convert(it) }
 }
 
 data class Mapping(
-    private val destinationRangeStart: Int,
-    private val sourceRangeStart: Int,
-    private val rangeLength: Int,
+    private val destinationRangeStart: Long,
+    private val sourceRangeStart: Long,
+    private val rangeLength: Long,
 ) {
-    private val sourceRange = IntRange(sourceRangeStart, sourceRangeStart + rangeLength - 1)
+    private val sourceRange = LongRange(sourceRangeStart, sourceRangeStart + rangeLength - 1)
 
-    fun find(sourceNumber: Int) =
+    fun find(sourceNumber: Long) =
         if (sourceRange.contains(sourceNumber))
             destinationRangeStart + (sourceNumber - sourceRangeStart)
         else
@@ -47,10 +47,10 @@ data class Mapping(
 }
 
 class Converter(private val mappings: List<Mapping>) {
-    fun convert(sourceNumber: Int) = mappings.firstNotNullOfOrNull { it.find(sourceNumber) } ?: sourceNumber
+    fun convert(sourceNumber: Long) = mappings.firstNotNullOfOrNull { it.find(sourceNumber) } ?: sourceNumber
 }
 
 class ConverterChain(private val converters: List<Converter>) {
-    fun convert(sourceNumber: Int) =
+    fun convert(sourceNumber: Long) =
         converters.fold(sourceNumber, { accumulator, converter -> converter.convert(accumulator) })
 }
